+++ reservation-service/src/main/java/kr/or/connect/jgb/service/CategoryService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import kr.or.connect.jgb.domain.Category;
 
 public interface CategoryService {
-	public Category get(Long id);
+	public Category get(int id);
 	public List<Category> getAll();
     public Category addCategory(Category category);
-    public int delete(Long id);
+    public int delete(int id);
     public int update(Category category);
 }
